+++ tests/run.js
@@ -10,6 +10,7 @@ import "./categories.test.js";
 import "./addons.test.js";
 import "./v02-features.test.js";
 import "./motion.test.js";
+import "./charts.test.js";
 import "./motion-editor.test.js";
 import "./security.test.js";
 import { run } from "./runner.js";